Move mock observation reader from mock package to observationtest.
